app/requirements.txt
cassandra-driver
venv-pack
pyarrow
fastparquet
pathvalidate
pyspark
tqdm
nltk